requirements.txt
streamlit
requests
python-docx
PyPDF2
reportlab
groq
google-generativeai
